+++ multidla/magicdla.h
@@ -1,7 +1,9 @@
 #ifndef MULTIDLA_MAGICDLA_H
 #define	MULTIDLA_MAGICDLA_H
 
-typedef short int FieldElement;
+#include <cstdint>
+
+typedef uint8_t FieldElement;
 
 const FieldElement FREE_CELL = 0;
 const FieldElement OCUPIED_CELL = 1;
